+++ src/main/java/oogasalad/model/players/AIPlayer.java
@@ -28,7 +28,7 @@ public class AIPlayer extends GenericPlayer{
 
   private List<Player> players;
 
-  public AIPlayer(Board board, int id, Map<String, Integer> inventory, int startingGold, Map<Integer, MarkerBoard> enemyMap) {
+  public AIPlayer(Board board, int id, Map<String, Double> inventory, int startingGold, Map<Integer, MarkerBoard> enemyMap) {
     super(board, id, inventory, startingGold, enemyMap);
     players = new ArrayList<>();
   }
